LayoutTests/http/tests/resourceLoadStatistics/aggregate-sorted-data-no-storage-access-database-expected.txt

Check correct UI data gets reported with no storage access.



Resource load statistics:

Registrable domain: subframe1
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    SubframeUnderTopFrameDomains:
        topframe1
    SubresourceUniqueRedirectsTo:
        topframe2
    isPrevalentResource: Yes
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: subframe2
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    SubframeUnderTopFrameDomains:
        topframe1
    isPrevalentResource: Yes
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: subframe3
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    SubframeUnderTopFrameDomains:
        topframe1
        topframe4
    SubresourceUnderTopFrameDomains:
        topframe3
    SubresourceUniqueRedirectsTo:
        topframe2
    isPrevalentResource: Yes
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: topframe1
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: topframe2
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: topframe3
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0
Registrable domain: topframe4
    hadUserInteraction: No
    mostRecentUserInteraction: -1
    grandfathered: No
    IsScheduledForAllButCookieDataRemoval: No
    isPrevalentResource: No
    isVeryPrevalentResource: No
    dataRecordsRemoved: 0

ITP Data:
Third Party Registrable Domain: subframe3
    {{ Has been granted storage access under topframe1: 0 },{ Has been granted storage access under topframe4: 0 },{ Has been granted storage access under topframe3: 0 },{ Has been granted storage access under topframe2: 0 },}
Third Party Registrable Domain: subframe1
    {{ Has been granted storage access under topframe1: 0 },{ Has been granted storage access under topframe2: 0 },}
Third Party Registrable Domain: subframe2
    {{ Has been granted storage access under topframe1: 0 },}
